+++ src/test/java/com/airbnb/metrics/KafkaStatsDReporterTest.java
@@ -11,6 +11,8 @@ import org.mockito.Matchers;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
+import java.util.Properties;
+
 import static org.mockito.Mockito.verify;
 
 public class KafkaStatsDReporterTest {
@@ -20,39 +22,20 @@ public class KafkaStatsDReporterTest {
   private StatsDClient statsD;
   private KafkaStatsDReporter reporter;
   private StatsDMetricsRegistry registry;
+  private Metric metric;
+  private final double value = 10.11;
 
   @Before
   public void init() throws Exception {
     MockitoAnnotations.initMocks(this);
     StatsDReporterConfig config = new StatsDReporterConfig(new VerifiableProperties());
     registry = new StatsDMetricsRegistry();
-    reporter = new KafkaStatsDReporter(
-        statsD,
-        registry,
-        config
-    );
-  }
+    reporter = new KafkaStatsDReporter(statsD, registry, config);
 
-  protected void addMetricAndRunReporter(
-      String metricName,
-      Metric metric,
-      String tag
-  ) throws Exception {
-    try {
-      registry.register(metricName, metric, tag);
-      reporter.run();
-    } finally {
-      reporter.shutdown();
-    }
-  }
-
-  @Test
-  public final void sendDoubleGauge() throws Exception {
-    final double value = 10.11;
-    Metric metric = new Metric() {
+    metric = new Metric() {
       @Override
       public MetricName metricName() {
-        return new MetricName("test-metric", "group");
+        return new MetricName("test-metric", "group", "description", "tagkey1", "valuekey1", "tagkey2", "valuekey2");
       }
 
       @Override
@@ -60,8 +43,31 @@ public class KafkaStatsDReporterTest {
         return value;
       }
     };
+  }
 
-    addMetricAndRunReporter("foo", metric, "bar");
-    verify(statsD).gauge(Matchers.eq("foo"), Matchers.eq(value), Matchers.eq("bar"));
+  protected void addMetricAndRunReporter(Metric metric) throws Exception {
+    try {
+      registry.register(metric);
+      reporter.run();
+    } finally {
+      reporter.shutdown();
+    }
+  }
+
+  @Test
+  public final void sendDoubleGaugeWithTagsEnabled() throws Exception {
+    addMetricAndRunReporter(metric);
+    verify(statsD).gauge(Matchers.eq("kafka.group.test-metric"), Matchers.eq(value), Matchers.eq("tagkey1:valuekey1,tagkey2:valuekey2"));
+  }
+
+  @Test
+  public final void sendDoubleGaugeWithTagsDisabled() throws Exception {
+    Properties properties = new Properties();
+    properties.put(StatsDReporterConfig.CONFIG_STATSD_TAG_ENABLED, "false");
+    StatsDReporterConfig config = new StatsDReporterConfig(new VerifiableProperties(properties));
+    reporter = new KafkaStatsDReporter(statsD, registry, config);
+
+    addMetricAndRunReporter(metric);
+    verify(statsD).gauge(Matchers.eq("kafka.group.tagkey1.valuekey1.tagkey2.valuekey2.test-metric"), Matchers.eq(value));
   }
 }